Splitting more modules and cleaning up
