+++ package.yaml
@@ -27,13 +27,15 @@ library:
     - &bytestring bytestring >=0.10
     - &split split >=0.2
     - &ShellCheck ShellCheck >=0.5.0
-    - gitrev >=1.3.1
     - aeson
     - text
     - bytestring
     - containers
     - void
     - mtl
+    - yaml
+    - filepath
+    - directory >= 1.3.0
 executables:
   hadolint:
     main: Main.hs
@@ -42,9 +44,6 @@ executables:
       - hadolint
       - optparse-applicative >= 0.14.0
       - gitrev >=1.3.1
-      - filepath
-      - directory >= 1.3.0
-      - yaml
       - text
       - containers
     when:
