+++ src/Hadolint.hs
@@ -1,0 +1,9 @@
+module Hadolint
+    ( module Hadolint.Lint
+    , module Hadolint.Rules
+    , module Hadolint.Config
+    ) where
+
+import Hadolint.Config
+import Hadolint.Lint
+import Hadolint.Rules
